[Protocol3] Cleaned up EdDSA code
