+++ packages/loopring_v3/test/eddsa.ts
@@ -2,19 +2,31 @@
 // https://github.com/iden3/circomlib
 
 const createBlakeHash = require("blake-hash");
+import crypto = require('crypto');
 const bigInt = require("snarkjs").bigInt;
 const babyJub = require("./babyjub");
 const poseidon = require("./poseidon.js");
+import { KeyPair, Signature } from "./types";
 
-exports.sign = sign;
-exports.verify = verify;
+export function getKeyPair() {
+    const entropy = crypto.randomBytes(32);
+    const secretKey = bigInt.leBuff2int(entropy).mod(babyJub.subOrder);
+    const publicKey = babyJub.mulPointEscalar(babyJub.Base8, secretKey);
 
-function sign(strKey, msg) {
+    const keyPair: KeyPair = {
+      publicKeyX: publicKey[0].toString(10),
+      publicKeyY: publicKey[1].toString(10),
+      secretKey: secretKey.toString(10),
+    };
+    return keyPair;
+}
+
+export function sign(strKey: string, msg: string) {
     const key = bigInt(strKey);
     const prv = bigInt.leInt2Buff(key, 32);
 
     const h1 = createBlakeHash("blake512").update(prv).digest();
-    const msgBuff = bigInt.leInt2Buff(msg, 32);
+    const msgBuff = bigInt.leInt2Buff(bigInt(msg), 32);
     const rBuff = createBlakeHash("blake512").update(Buffer.concat([h1.slice(32,64), msgBuff])).digest();
     let r = bigInt.leBuff2int(rBuff);
     r = r.mod(babyJub.subOrder);
@@ -25,29 +37,27 @@ function sign(strKey, msg) {
     const hasher = poseidon.createHash(6, 6, 52);
     const hm = hasher([R8[0], R8[1], A[0], A[1], msg]);
     const S = r.add(hm.mul(key)).mod(babyJub.subOrder);
-    return {
-        R: R8,
-        S: S,
-        hash: hm.toString(10),
+
+    const signature: Signature = {
+      Rx: R8[0].toString(),
+      Ry: R8[1].toString(),
+      s: S.toString(),
     };
+    return signature;
 }
 
-function verify(msg, sig, pubKey) {
+export function verify(msg: string, sig: Signature, pubKey: string[]) {
     const A = [bigInt(pubKey[0]), bigInt(pubKey[1])];
     const R = [bigInt(sig.Rx), bigInt(sig.Ry)];
     const S = bigInt(sig.s);
 
     // Check parameters
-    if (!Array.isArray(R)) return false;
-    if (R.length!= 2) return false;
     if (!babyJub.inCurve(R)) return false;
-    if (!Array.isArray(A)) return false;
-    if (A.length!= 2) return false;
     if (!babyJub.inCurve(A)) return false;
-    if (S>= babyJub.subOrder) return false;
+    if (S >= babyJub.subOrder) return false;
 
     const hasher = poseidon.createHash(6, 6, 52);
-    const hm = hasher([R[0], R[1], A[0], A[1], msg]);
+    const hm = hasher([R[0], R[1], A[0], A[1], bigInt(msg)]);
 
     const Pleft = babyJub.mulPointEscalar(babyJub.Base8, S);
     let Pright = babyJub.mulPointEscalar(A, hm);
